fix kafkaUser and kafkaPW nullpoint exception

+++ test/hongquan_format_big_endian.sh
@@ -59,6 +59,7 @@ javac -d bin -cp example/:libs/* -Xlint:deprecation example/ProducerTest.java
 java -cp bin:bin/*:libs/* ProducerTest
 
 cd $KAFKA_CDC/bin
+
 ./KafkaCDC-server.sh -p $PARTITION  -b $BROKER -d $DBIP -s $DESTSCHEMA --table $TABLE -t $TOPIC -f HongQuan --bigendian --full start --sto 5 --interval 2 --key org.apache.kafka.common.serialization.LongDeserializer --value org.apache.kafka.common.serialization.ByteArrayDeserializer
 
 #get result file from $USER
@@ -140,5 +141,7 @@ RESULT="$currentTime hongquan_format_big_endian failed"
 fi
 
 $KAFKA/bin/kafka-topics.sh --delete --zookeeper $ZOOKEEPER --topic $TOPIC
-rm -f $RESULTPATH
-rm -f $EXPECTPATH
+if [ "x${DEBUG}" != "xYES" ]; then
+  rm -f $RESULTPATH
+  rm -f $EXPECTPATH
+fi
