+++ test/hongquan_format_fix_types.sh
@@ -17,6 +17,7 @@ EXPECTDIR="$EXPECTDIR"
 FINALRESULTPATH="$FINALRESULTPATH"
 RESULTPATH="$EXPECTDIR/${TOPIC}_result.log"
 EXPECTPATH="$EXPECTDIR/${TOPIC}_expect.log"
+DEBUG=$1
 
 expect <<-EOF
   set timeout 60
@@ -136,5 +137,7 @@ else
 fi
 
 $KAFKA/bin/kafka-topics.sh --delete --zookeeper $ZOOKEEPER --topic $TOPIC
-rm -f $RESULTPATH
-rm -f $EXPECTPATH
+if [ "x${DEBUG}" != "xYES" ]; then
+  rm -f $RESULTPATH
+  rm -f $EXPECTPATH
+fi
