+++ test/hongquan_format_varchar_type.sh
@@ -139,5 +139,7 @@ else
 fi
 
 $KAFKA/bin/kafka-topics.sh --delete --zookeeper $ZOOKEEPER --topic $TOPIC
-rm -f $RESULTPATH
-rm -f $EXPECTPATH
+if [ "x${DEBUG}" != "xYES" ]; then
+  rm -f $RESULTPATH
+  rm -f $EXPECTPATH
+fi
