+++ test/multikey_commit_finally.sh
@@ -150,7 +150,9 @@ RESULT="$currentTime ${TOPIC} failed"
 fi
 
 $KAFKA/bin/kafka-topics.sh --delete --zookeeper $ZOOKEEPER --topic $TOPIC
-rm -f $DATAFILE
-rm -f $RESULTPATH
-rm -f $EXPECTPATH
+if [ "x${DEBUG}" != "xYES" ]; then
+  rm -f $DATAFILE
+  rm -f $RESULTPATH
+  rm -f $EXPECTPATH
+fi
