+++ test/update_insert_row_finally.sh
@@ -144,6 +144,8 @@ else
 fi
 
 $KAFKA/bin/kafka-topics.sh --delete --zookeeper $ZOOKEEPER --topic $TOPIC
-rm -f $DATAFILE
-rm -f $RESULTPATH
-rm -f $EXPECTPATH
+if [ "x${DEBUG}" != "xYES" ]; then
+  rm -f $DATAFILE
+  rm -f $RESULTPATH
+  rm -f $EXPECTPATH
+fi
